requirements.txt
Flask==3.0.0
youtube-transcript-api==1.2.2
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.1.1